move validateTitle & isSlugUnique fn to moduleStore
