+++ src/pages/Howto/Content/CreateHowto/CreateHowto.tsx
@@ -98,7 +98,7 @@ export class CreateHowto extends React.Component<IProps, IState> {
   }
 
   public validateTitle = async (value: any, meta?: FieldState) => {
-    this.store.validateTitle(value, meta)
+    this.store.validateTitle(value, 'howtosV1', meta)
   }
 
   // automatically generate the slug when the title changes
